minor change (unsigned int --> unsigned long) for addresses

+++ pru_sw/app_loader/interface/prussdrv.c
@@ -689,19 +689,19 @@ void *prussdrv_get_virt_addr(unsigned int phyaddr)
         && (phyaddr <
             prussdrv.pru0_dataram_phy_base + prussdrv.pruss_map_size)) {
         address =
-            (void *) ((unsigned int) prussdrv.pru0_dataram_base +
+            (void *) ((unsigned long) prussdrv.pru0_dataram_base +
                       (phyaddr - prussdrv.pru0_dataram_phy_base));
     } else if ((phyaddr >= prussdrv.l3ram_phys_base)
                && (phyaddr <
                    prussdrv.l3ram_phys_base + prussdrv.l3ram_map_size)) {
         address =
-            (void *) ((unsigned int) prussdrv.l3ram_base +
+            (void *) ((unsigned long) prussdrv.l3ram_base +
                       (phyaddr - prussdrv.l3ram_phys_base));
     } else if ((phyaddr >= prussdrv.extram_phys_base)
                && (phyaddr <
                    prussdrv.extram_phys_base + prussdrv.extram_map_size)) {
         address =
-            (void *) ((unsigned int) prussdrv.extram_base +
+            (void *) ((unsigned long) prussdrv.extram_base +
                       (phyaddr - prussdrv.extram_phys_base));
     }
     return address;
